Fixing naming mismatch between documentation and code, addressing Issue https://github.com/SE-Stuttgart/moodle-block_slidefinder/issues/16.
Also fixed a mistake where only the userid given by the webservice was checked by the system for course access, and not the webservice itself. It now checks both. It is now clear that both the webservice user (as mentioned in Issue https://github.com/SE-Stuttgart/moodle-block_slidefinder/issues/17) and the provided userid (intended for the user calling upon the webservice) need to have access to the course.

Small update to the web service description, addressing Issue https://github.com/SE-Stuttgart/moodle-block_slidefinder/issues/17.